Доработка валидации формы загрузки изображения
