+++ js/load-form-entities.js
@@ -1,0 +1,6 @@
+export const loadNewButton = document.querySelector('.img-upload__input');
+export const loadNewPopup = document.querySelector('.img-upload__overlay');
+export const imageLoadForm = /** @type {HTMLFormElement} */(document.querySelector('.img-upload__form'));
+export const hashTagInput = /** @type {HTMLInputElement} */(imageLoadForm.querySelector('.text__hashtags'));
+export const commentInput = /** @type {HTMLTextAreaElement} */(imageLoadForm.querySelector('.text__description'));
+export const submitButton = /** @type {HTMLButtonElement} */(imageLoadForm.querySelector('.img-upload__submit'));
